Use production site URL when building from master branch (#2)
If we're building from the master branch, use the production URL and not the unique deploy one.

When developing we used the site URL that's unique to each deploy. But this was done on all deploys including from master branch. This means the production URL (cyberqueer.netlify.com) would not work because all the code expects it to be at a different domain.